Update ProductsGroupsController.php
Changes to save and restore the user specified sort order for the "Product Group"

actionProducts() :
When displaying the items, we need to use the sort order from the database to present the products in the correct order to the user.

actionProductsUpdate():
When Saving the group we are using the sort order provided by the user rather than calculating it.   Also clear the sort order from any products which are no longer in the group.